Remove bits from stock license file
Considering this includes some image assets from the original launcher, we can't claim public domain because we don't own those assets. This is being released without a license considering it is primarily a modification of reverse engineered code.

In the event someone doesn't like this being so close to the original, I can write a brand new replacement later if it comes down to it.